% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berry_logit.R
\name{berry_logit_data}
\alias{berry_logit_data}
\title{Prepare a data set with product data for estimation of a simple logit model
without random coefficients as initially described in Berry (1994)}
\usage{
berry_logit_data(productData, lhs = "log_s_log_s0", s = NULL, s0 = NULL,
  market_identifier = NULL)
}
\description{
Prepare a data set with product data for estimation of a simple logit model
without random coefficients as initially described in Berry (1994)
}
