% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berry_logit.R
\name{berry_logit_get_elasticities}
\alias{berry_logit_get_elasticities}
\title{Compute elasticities for the simple logit model without
random coefficients initially described in Berry (1994)}
\usage{
berry_logit_get_elasticities(productData, estimation, variable, product_id,
  market_identifier, s = as.character(formula(estimation)[[2]]),
  products = NULL, markets = NULL, simplify = TRUE)
}
\description{
Let s_i be the market share of firm i, p_i the price,
and beta_b the coeficient of price in the utility function.
}
\details{
The own price derivative is

ds_i / dp_i = -beta_p * s_i * (1-s_i)

The own price elasticity is

ds_i / dp_i * (p_i / s_i) = -beta_p * (1- s_i) * p_i

The cross price derivative with respect to the price of a good j is given by

ds_i / dp_j = beta_p * s_i * s_j

The cross price elasticity is

ds_i / dp_j * (p_j / s_i) = beta_p * s_j * p_j
}
