% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blp_tools.R
\name{blp_create_instruments}
\alias{blp_create_instruments}
\title{Compute BLP instruments and add to product data}
\usage{
blp_create_instruments(productData, vars, market_identifier, firm_var = NULL,
  prefix = NULL, add_cols = TRUE)
}
\arguments{
\item{productData}{The data frame that contains the product data}

\item{vars}{The variable names of all the product attributes for which instruments shall be computed}

\item{market_identifier}{The variable name that identifies a market}

\item{firm_var}{The variable that identifies a firm. If NULL only a single instrumental variable per attribute is computed that consists of the sum of the attribute for all other products.}

\item{prefix}{a prefix to add to the new instrumental variables. If NULL a default will be chosen. If firm_var is not NULL prefix must be a vector of two strings. The first is the prefix for the intra-firm sums and the second for the inter-firm sums.}

\item{add_cols}{if TRUE the new instrumental variables will be added as new columns to product data. Otherwise only a data frame with the new instruments will be returned.}
}
\value{
A data frame that contains the new instruments as columns. The data frame also has an attribute "instruments" that contains the names of all generated instrumental variables.
}
\description{
BLP suggest to instrument prices with the values of other product
characteristics.
}
\details{
If the characteristics of competing products change between markets,
one may react with the price for the own product. For example, if ceteris
paribus other products get more attractive attributes, it can be optimal
to reduce the prices of the own product to mitigate market share losses.

The validity of these instrument depends on the stronger assumption that
other firm's observed attributes are not correlated with the own product's
unobserved attributes.

Consider a BLP model for the demand of cars and the attribute horse power.
Concrentely BLP suggest to compute to instrumental variables
related to this attribute:

1. The sum of horse power of all other cars from the same firm

2. The sum of horse power of all cars from a different firms

This function allows to quickly compute such instrumental variables
}
