% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blp_tools.R
\name{blp_get_derivatives}
\alias{blp_get_derivatives}
\title{Similar to get_elasticities, but gets corresponding derivatives}
\usage{
blp_get_derivatives(blp_data, blp_estimation, variable, products = NULL,
  market, e = NULL)
}
\description{
E.g. if the variable is the price, we compute the matrix
of cross-price derivatives for the selected products for
a particular market.
}
\details{
This matrix is crucial to compute Bertrand equilibria.
}
