% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blp_tools.R
\name{blp_make_formula}
\alias{blp_make_formula}
\title{Helper function to generate a formula for}
\usage{
blp_make_formula(dep, lin = NULL, exo = NULL, rc = NULL, inst = NULL)
}
\description{
Helper function to generate a formula for
}
